+++ src/views/UserDetail/UserDetail.js
@@ -33,7 +33,6 @@ import {
 } from '@folio/stripes/components';
 import {
   NotesSmartAccordion,
-  ViewCustomFieldsRecord,
   NotePopupModal,
 } from '@folio/stripes/smart-components';
 
@@ -51,6 +50,7 @@ import {
   UserAffiliations,
   UserServicePoints,
   ReadingRoomAccess,
+  ViewCustomFieldsSection,
 } from '../../components/UserDetailSections';
 import HelperApp from '../../components/HelperApp';
 import IfConsortium from '../../components/IfConsortium';
@@ -76,8 +76,13 @@ import ActionMenuDeleteButton from './components/ActionMenuDeleteButton';
 import OpenTransactionModal from './components/OpenTransactionModal';
 import DeleteUserModal from './components/DeleteUserModal';
 import ExportFeesFinesReportButton from './components';
+import { CUSTOM_FIELDS_SECTION } from '../../constants';
 
 import css from './UserDetail.css';
+
+export const ACCORDION_ID = {
+  CUSTOM_FIELDS: 'customFields',
+};
 
 class UserDetail extends React.Component {
   static propTypes = {
@@ -207,7 +212,7 @@ class UserDetail extends React.Component {
         rolesSection: false,
         servicePointsSection: false,
         notesAccordion: false,
-        customFields: false,
+        [ACCORDION_ID.CUSTOM_FIELDS]: false,
         readingRoomAccessSection: false,
       },
     };
@@ -691,7 +696,7 @@ class UserDetail extends React.Component {
       );
     }
 
-    if (!user) {
+    if (!user || resources.selUser.isPending) {
       return (
         <LoadingPane
           id="pane-userdetails"
@@ -755,6 +760,7 @@ class UserDetail extends React.Component {
                   settings={settings}
                   stripes={stripes}
                   expanded={sections.userInformationSection}
+                  customFields={customFields}
                   onToggle={this.handleSectionToggle}
                 />
 
@@ -796,6 +802,7 @@ class UserDetail extends React.Component {
                   onToggle={this.handleSectionToggle}
                   departments={departments}
                   userDepartments={userDepartments}
+                  customFields={customFields}
                 />
                 <ContactInfo
                   accordionId="contactInfoSection"
@@ -804,17 +811,15 @@ class UserDetail extends React.Component {
                   addresses={addressesList}
                   addressTypes={addressTypes}
                   expanded={sections.contactInfoSection}
+                  customFields={customFields}
                   onToggle={this.handleSectionToggle}
                 />
-                <ViewCustomFieldsRecord
-                  accordionId="customFields"
+                <ViewCustomFieldsSection
+                  accordionId={ACCORDION_ID.CUSTOM_FIELDS}
+                  customFields={customFields}
+                  expanded={sections[ACCORDION_ID.CUSTOM_FIELDS]}
+                  sectionId={CUSTOM_FIELDS_SECTION.CUSTOM_FIELDS}
                   onToggle={this.handleSectionToggle}
-                  expanded={sections.customFields}
-                  backendModuleName="users"
-                  entityType="user"
-                  customFieldsValues={customFields}
-                  customFieldsLabel={<FormattedMessage id="ui-users.custom.customFields" />}
-                  noCustomFieldsFoundLabel={<FormattedMessage id="ui-users.custom.noCustomFieldsFound" />}
                 />
                 {
                   displayReadingRoomAccessAccordion && (